Demo object of Parallax modified to solve the fact that a re-power of the module
In manyof our battery operated robot applications we want to use a gyro to measure angular speed. The Parallax 27911gyro is a good one. But as the unit seems to work on supply voltages as long as 2.5V the unit often ends up half way a power down and reboot if the supply voltage does not drop below 2V. In a 5V/3.3.V logic supply, this may take a minute or longer. So to everyones surprise the gyro shows non sense values after re-powering the control system. This has been solved by adding a reboot command in the gyro object